cros lint: require encoding='utf-8' with pathlib open calls
When opening a file in text mode, the default encoding used is whatever
the system default is which can be just about anything.  To avoid errors
when run in non-UTF-8 locales, force people to always specify encoding
and be set to UTF-8.

This also helps remind people when they want to open files in text or in
binary mode.  Ad-hoc & unittest coverage isn't always sufficient.

For now we only check pathlib.open calls.  This specific API is not used
a lot, so it's easy to flip on.  We can add more APIs on top of this.

BUG=b:187789896
TEST=`cros lint` is clean in chromite

Change-Id: I436bbfedafa304d306bd2e62153a2507b8864c62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2744859
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Commit-Queue: Sloan Johnson <sloanjohnson@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>