service: do BundleImageZip with 7z
This is a faster method of producing the same file, we set the algorithm
to DEFLATE instead of STORE, and it produces a ZIP64 compatible output. This should be ok for nearly all clients newer than ~winxp.

Mini-local-benchmark puts this as twice as fast. (2.5 minutes from 7
minutes.)

BUG=None
TEST=./run_tests

Change-Id: I8451da1507ed55af7c8e9c0897a487737189e37e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3956725
Commit-Queue: George Engelbrecht <engeg@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: George Engelbrecht <engeg@google.com>