cli: clean-outdated-pkgs improve package filtering
cros clean-outdated-pkgs has an optimization for no-op case: if no
packages were deleted in phase 1 (old pkgs with no ebuild),
then we don't do phase 2 (slot conflict search).
However, previously, if we found an old package that was not deleted
due to being in the DO-NOT-PURGE list, the script still went into phase
2.
This change fixes it.

BUG=b:251829027
TEST=manual

Change-Id: I52054b922ad4fa824bb0483f64e4b74a13172a9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4137169
Tested-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sergey Frolov <sfrolov@google.com>