lib: image_lib: use integers for partition values
The parted parsing code returns floats for the partition values.
This is because the code used to support different units beyond
bytes -- you could get KB, MB, GB, etc..., and using floats for
those makes sense (since partitions start/size are not required
to be aligned to such values).  The unit option was dropped a
long time ago in favor of always returning bytes, but the float
format was retained.

This is problematic because the cgpt parsing code only returns
integers, never floats, and two are supposed to behave the same
so as to be transparent to the caller.  So convert the parted
logic to only use integers.

BUG=b:256896261
TEST=CQ passes

Change-Id: I4297b5bb0eb4fab4a9226bd890db7c600d99b030
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4150007
Commit-Queue: Ram Chandrasekar <rchandrasekar@google.com>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>