lib: image_lib: release loopback device on attach failure
If, after we've actually attached the image to a loopback, we crash,
we leak the device because we never communicate back up the name of
the device that was allocated.  Change our code to release it instead
whenever we crash since there isn't much else we could usefully do.

BUG=b:256896261
TEST=CQ passes

Change-Id: I15cd572bb1ca8cedab2c177fe8b5d6e4144bf75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4167183
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>