android: detect runtime artifacts pin change in WriteLKGB
Android uprevs are only triggered when new Android builds are published.
However on release branches, we need to trigger an uprev as well when
the runtime artifacts pin is modified.

This CL introduces a field in the LKGB file generated by WriteLKGB, such
that a modification is made when the runtime artifacts pin changes, and
in turn triggers an uprev.

Notes when cherry-picking: this CL works with M110+ branches; for older
branches, the runtime artifacts pin is at a different location.

BUG=b:258558620
TEST=./run_tests
TEST=./api/contrib/call_scripts/android__write_lkgb
TEST=led run https://ci.chromium.org/swarming/task/5fdde541d9f71010?server=chromeos-swarming.appspot.com

Change-Id: Ic2f22b63883da4ed99ddb311fd9448c3c1fc3846
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4173177
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Reviewed-by: Yury Khmel <khmel@google.com>
Commit-Queue: Shao-Chuan Lee <shaochuan@chromium.org>
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>