cros lint: require encoding='utf-8' with pathlib write_text calls
When writing a file in text mode, the default encoding used is whatever
the system default is which can be just about anything.  To avoid errors
when run in non-UTF-8 locales, force people to always specify encoding
and be set to UTF-8.

This also helps remind people when they want to write files in text or
in binary mode.  Ad-hoc & unittest coverage isn't always sufficient.

BUG=b:187789896
TEST=`cros lint` is clean in chromite

Change-Id: Icd7f845d20490ce2f0a465db48894cbc08ca0657
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4188829
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>