cros lint: require encoding='utf-8' with open calls
When opening a file in text mode, the default encoding used is whatever
the system default is which can be just about anything.  To avoid errors
when run in non-UTF-8 locales, force people to always specify encoding
and be set to UTF-8.

This also helps remind people when they want to open files in text or in
binary mode.  Ad-hoc & unittest coverage isn't always sufficient.

BUG=b:187789896
TEST=`cros lint` is clean in chromite

Change-Id: If1df36847ee13cf6c7254537db3e11efd5251ca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4196514
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>