lib: git: Add a ls-files wrapper
This is useful to the follow-on CL, so that we can get relevant (not
gitignore'd) files to format.

git ls-files has a ton of options and behaviors.  I went with the
YAGNI approach here of only implementing what we need right now, and
if someone wants to go extend this wrapper later to support some of
those flags, it shouldn't be difficult.

BUG=b:266643447
TEST=unit test provided

Change-Id: I5d3d53906690f4fb73cabe1e1dd937776d0bcd79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4190174
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>