cli: format: Accept a directory as input
Conventionally, most formatting tools will accept a directory as
input, and recursively format all files in that directory.  For
example, formatting an entire directory with black is usually as
simple as running "black .".

Users may want the same convenience in cros format.  When we're in a
git repository, they likely also want the .gitignore to be
respected (and not formatted).

BUG=b:266643447
TEST=provided unit tests

Change-Id: I62f7d488268930aae034b74a48b855b0ec173bc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4190175
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>