cros lint: handle variable arguments to open calls
The vast majority of open calls use constants for mode & encoding.
But if a variable is used, the current linter code crashes because
it tries to read a .value which doesn't exist.  Gracefully handle
this by ignoring it since we can't know the value.

BUG=b:187789896
TEST=`cros lint` doesn't crash with some chromite code

Change-Id: I7ce9d4cdfe9df1522d642b1b5b9db08b3519bcae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4196594
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>