lib: Copy miniOS kernel into A+B partitions
.. because images should now allocate enough space for both A + B
partitions of miniOS.
Even if the copies of the kernel are initially the same during build,
that's fine. Scripts that sign the image later can decide how to sign
each, whether that's with same keys or different is not up to the flow
here.

BUG=b:267545513
TEST=./run_tests

Change-Id: Iaeea6360c9963e367dd85f46285c428ac89643e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4229283
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>