format: switch to lazy module loading
As we add more formatters, the number of modules we load grows, and
this slows down most invocations of cros format.  For example, if we
assume that most use of cros format only touches 1 or 2 file types
at a time, that's only ~10% of the current set of formats, which
means 90% of the formatter modules are useless.

Concretely, this takes `cros format --help` from ~210ms to ~190ms.
Not massive, but it can add up, and as mentioned above, this gap
will only grow -- lazy loading is constant.

BUG=None
TEST=CQ passes

Change-Id: I74d06b5a49392144d1e87b34c0fe033fd12faf12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4248777
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>