scripts: crosfw: Reconfigure automatically if needed
If anything has changed in Kconfig or the defconfig files, force a
reconfigure. This avoid build errors or building with the wrong config.

BUG=b:268384689
TEST=crosfw sandbox
Change a Kconfig and try again, see that it says it is reconfiguring
Do the same with configs/sandbox_spl_defconfig

Change-Id: I28590ab318a6e1090e93d77c09a06f5b9089ae7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4248403
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>