scripts: package rustc and cargo
These were probably dropped last year, when we split `rust` and
`rust-host` into separate packages. It seems this broke a user of them.

We don't officially support the functionality of these, but we have
promised to ship them as a part of this tarball; this CL helps us make
good on that promise.

BUG=b:269306499
TEST=cros_setup_toolchains produces tarballs with these. Running
     `./rustc --help` and `./cargo --help` works.

Change-Id: I0b3c160dd1388bf16dcf459ca1d85c7a7bf91311
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4249801
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>