scripts: crosfw: Allowing building in-tree
Sometimes it is useful to build in the source tree, either to check that
it still works OK, or because only one board is being worked on.

Add an option for this.

Change mention of distclean to mrproper since that is what the U-Boot
build system advises.

BUG=b:268384689
TEST=crosfw -I sandbox
$ ./u-boot -Dl
See that the build completed OK

Change-Id: Ifab4f72351879647d20feab7ff5639455aa7487c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4255446
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>