scripts: crosfw: Allow providing an external devicetree
In some cases we want to provide a pre-build devicetree instead of
using the one within U-Boot, e.g. if it is coming from another project.
Add an option to support this.

BUG=b:268384689
TEST=crosfw -L sandbox --dtb /tmp/b/snow/u-boot.dtb -v
See that 'EXT_DTB=/tmp/b/snow/u-boot.dtb' is shown in the build command

Change-Id: Idcca8a22bbc3bec463c47a86e1fc00a76c412dc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4263382
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>