scripts: crosfw: Avoid hanging with invalid defconfig
If a board does not provide a value for an integer Kconfig option,
the conf tool will normally prompt for it. This hangs forever since our
script does not provide input.

This is actually always a bug in the defconfig, so use the existing
KCONFIG_NOSILENTUPDATE feature to detect it and print a proper error.

BUG=b:268384689
TEST=Comment out CONFIG_BOOTSTAGE_STASH_ADDR in sandbox_defconfig
$ crosfw -L sandbox
See that it now shows a helpful error instead of hanging:

   .config:43:warning: symbol value '' invalid for BOOTSTAGE_STASH_ADDR
    *** The configuration requires explicit update.

Change-Id: Ia60751b7a1abe433830a2f135a782a11a938b415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4269692
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>