scripts: Add key to remote latest SDK file
In the new recipes-based SDK builder, we don't want to update the latest
SDK as soon as we build it, since the uprev/test steps come later. Thus,
we need to create a way to distinguish between the latest SDK for users
and the next SDK uprev target.

This CL just changes the format of the remote latest file. In a future
CL I'll add onto it so that `upload_prebuilts.py` takes in a CLI flag
specifying which of the two keys to update. The other key should retain
the existing value.

The unittest change here is pretty mild, but again, it'll be more
significant once the script has new logic.

BUG=b:270142110
TEST=./run_tests

Change-Id: If5854e39a11ca1f14364dc3bfa8f9b9a465f4399
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4283093
Commit-Queue: Greg Edelston <gredelston@google.com>
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Cindy Lin <xcl@google.com>