lib: cros_sdk_lib: Provide way to phase out depot_tools from PATH
For a very long time, we've kept depot_tools on $PATH in the SDK. This
provides tools like vpython3, which lots of tools have grown to rely on;
but it also contains a bundle of other tools of varying utility within
the SDK. We don't really want them all in the PATH, and some (like
`repo`) actively create confusion, when users don't know whether they
should run them in the SDK or not.

Let's provide a way to control which tools are available in the SDK,
rather than take them all-or-nothing.

Note: this only affects newly-created chroots. We'll have a chroot
upgrade hook for existing ones.

BUG=b:270994985
TEST=./run_tests

Cq-Depend: chromium:4297135
Change-Id: Ib20173fbb84ec3bc1ecfcf45b6d1cb511e8198f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4297138
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>