scripts: crosfw: Show warnings correctly
At present stderr is suppressed if it does not cause a build error.
Most of the time we don't want to see stdout, but we do want the
warnings.

Rather than redirecting stderr to stdout, show it so that warnings
are visible.

BUG=b:268384689
TEST=crosfw -L sandbox    (on something with warnings)
See that they appear on the console and in an IDE

Change-Id: Iefa8c8bd05606d8e990430ad7eb3f880e9069b64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4299816
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>