scripts: crosfw: Show warnings correctly
At present stderr is suppressed if it does not cause a build error.
Most of the time we don't want to see stdout, but we do want the
warnings.

Rather than redirecting stderr to stdout, show it so that warnings
are visible.

BUG=b:268384689
TEST=crosfw -L sandbox    (on something with warnings)
See that they appear on the console and in an IDE

Change-Id: Iefa8c8bd05606d8e990430ad7eb3f880e9069b64