lib: fix a bug in LoadKeyValueStore
Running `strace` on a silly test of LoadKeyValueStore reveals that
`gsutil` doesn't write directly to the requested file. It instead writes
to a tempfile ending in `_.gstmp`, and renames over the requested file.
This means that we need to reopen the file, as our `NamedTemporaryFile`
is empty and nameless. (...I feel bad for the file somehow after writing
that)

BUG=b:271306977
TEST=Ran this on `gs://chromiumos-sdk/cros-sdk-latest.conf`.

Change-Id: I4e80bacf99dbaeddec9f805569b3cb1b7a7d9b17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4311846
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>