sdk: bin: No symlinks; just exec
For $reasons, we can't rely on depot_tools executables operating
correctly when run via symlink. Rewrite these symlinks as shell scripts,
with:

  for s in *
  do
    rm "$s"
    cat <<"EOF" >$s
  #!/bin/sh
  exec /mnt/host/depot_tools/"$(basename "$0")" "$@"
  EOF
  done
  chmod 755 *

Because these are so short, and because I don't want to pollute this
path (which will be on $PATH) with another executable helper, I
duplicate the "logic" into each of these. Most of these will be removed
shortly anyway, ideally.

BUG=b:270994985
TEST=run these tools, once they're on $PATH in SDK

Change-Id: Id0a76265b8699cb6f8b4bc81c49d9aefdefca8b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4313108
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>