lib: skyrim: Remove dut_control_off delay
Remove the 1 second delay that exists before resetting the AP.
It was added to workaround an issue on the Ti50 where it wasn't
releasing SPI bus to the AP quickly enough.

A proper fix on the Ti50 has been released so we don't need the
workaround anymore.

BUG=b:245777966
TEST=cros ap flash -b skyrim -i <image>
     Observe that the AP boots after flashing

Change-Id: I456ec0371969d5954856dde6e0aedbb1fee8004d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4328471
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Robert Zieba <robertzieba@google.com>
Auto-Submit: Mark Hasemeyer <markhas@google.com>
Tested-by: Mark Hasemeyer <markhas@google.com>