lib: ap_firmware_config: skyrim.py: delete servo.is_micro case
This is no-op cleanup.  Skyrim has C2D2, not servo header.
Skyrim never had a servo header, nor did it have the
daughterboard with servo header that this "servo.is_micro" case
was meant for.  That was copied from guybrush, and never applied
to skyrim, and even for guybrush it's obsolete now.

BUG=b:253349276
BUG=b:258719429
TEST=Compared before and after processed config from command:
$ cros ap dump-config -b skyrim
The only diffs are removal of servo_micro cases.

Change-Id: Ifd51254606dfbf36727abb84e479bf2848dbf543
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4342752
Tested-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Mark Hasemeyer <markhas@google.com>
Commit-Queue: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>