cli: cros_chrome_sdk: disable HEVC arg on non-internal builds
We changed the way we enable the HEVC codec, so now the simple chrome
workflow for boards w/ HEVC enabled that don't use the --internal flag
to the Chrome SDK fail to build with an assertion error. This unsets the
HEVC arg for non-internal builds since there are other dependencies that
need to be satisfied that aren't.

BUG=b:272127508
TEST=CQ

Change-Id: I969aa8e940d821325eacaa1feb445a6a27b62df6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4347506
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>