scripts: Don't fail if no remote SDK uprev target
For some reason, the current remote SDK latest file does not list an SDK
latest uprev target. I don't know why, and we need to figure that out.
Regardless, when we're just modifying the LATEST_SDK, we don't need to
fail if the latest uprev target is not found. Instead, set it to None,
which isn't great but it doesn't decrease information and it's better
than failing.

This should fix the attached bug, which has caused the (legacy) SDK
builder to fail for the past 4 days.

BUG=b:274196697
TEST=run_tests

Change-Id: If447a87877910973d7c14c66790c7dbabfd3aa24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4355315
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: Cindy Lin <xcl@google.com>
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>