lib: toolchain_util: Profile type in orderfile name
Replace "arch" with "profile" in the orderfile name.
"profile" is passed in profile_info["chrome_cwp_profile"] and currently
matches with "arch" in the orderfile builders so this renaming doesn't
affect the current workflow.

In the future we might want to try experimental orderfiles and changing
"profile" name will be the way to distinguish them from the production
profiles.

pyright kindly pointed out at some type inconsistencies which I also
addressed in this CL.

BUG=b:262734900
TEST=unit tests

Change-Id: Ic91af643034fa12d4115b3d1cd32c204e3bf257a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4359052
Tested-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>