cli: cros: lint/format: move no-op message until after processing
Due to some refactors, the code has ended up in a weird state where
we assign files to the user inputs, and then after checking them,
turn around and clobber the values.  Delay the check until after we
process the user inputs to avoid this and to be a bit more clear as
to what's happening.

BUG=None
TEST=`cros lint` & `cros format` still warn & exit 0

Change-Id: I58e37b8ae7c81f9ca368572f4a6f2eff112a4448
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4362866
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Commit-Queue: Ram Chandrasekar <rchandrasekar@google.com>
Auto-Submit: Mike Frysinger <vapier@chromium.org>