cros-ide: Efficiently update Gerrit comment threads
Before this change when comments are updated on Gerrit the extension
would clear all the comment threads and recreate the threads to display
from scratch. After this change the extension utilizes existing threads
to achieve the desired end state. This is computationally more efficient
and less distracting to the user.

This change also fundamentally fixes the issue that the same comment
duplicates, because now we atomically update the data structure managing
vscode comment threads.

BUG=b:270101085
BUG=b:273168590
FIXED=b:273168590
TEST=npm run unit-test
TEST=manually install the extension and use it for a while.

Change-Id: I1a75334f83f652b34bbe5f51c74a70b056a2281e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4368392
Reviewed-by: Eric Lok <lokeric@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>