contrib: Add subcmd to downstreaming script to rm self from attn set
Sometimes after running the downstreaming script, the user does not get
removed from the attention sets of the CLs after they merge. This leaves
the user's dashboard cluttered with dozens of CLs that no longer need
attention and are tedious to remove.

This CL adds a subcommand parser to the script along with
an inaugural `clear_attention` subcommand that removes the current user
from all merged CLs under the `zephyr-downstream` topic. The original,
core functionality of the downstreaming script is unaffected (no CLI
changes for normal downstreaming).

BUG=None
TEST=./zephyr_downstream clear_attention

Change-Id: I9924f6012f6950e5fa5c96ae19e3ed93d1acb0aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4374330
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>