lib: osutils: Move: Remove nested/overlaid "from" dirs
With MoveDirContents(allow_nonempty=True), we might be iterating into a
subdirectory to move its contents over to a similar directory on the
destination. Unfortunately, we don't remove the subdirectory in question
after we've iterated through it.

In other words, we fail this piece of the new unit test:

    self.assertNotExists(self.from_dir / "a")

BUG=none
TEST=./run_tests

Change-Id: Ied17808b6f64c40506c0f6c93fb0e2a6f8f41243
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4382215
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Gilberto Contreras <gcontreras@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>