lib: sdk: Migrate chroot/tmp path to out/tmp
Users should be using chroot_lib and/or path_util for path translation,
so this should [*] be seamless.

Note to sheriffs, etc.: If we see issues around not finding artifacts
that should be in /tmp from outside the chroot, this CL may be at fault.
It should be safe to revert, although that will likely leave some split
between old artifacts in chroot/tmp vs. out/tmp. (That's safe to do,
although perhaps confusing to the average user.)

[*] Famous last words.

BUG=b:265885353
TEST=./run_tests
TEST=`cros_sdk` moves existing chroot/tmp files to out/tmp; (also, with
     --chroot and --out-dir args)

Change-Id: I76d853d53222b33909df780f7b31bcc361245982
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4391052
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>