lib: build_query: Beginnings of this module
Start building out lib/build_query.py, which is the Python
implementation and API for the "cros query" command (CLI front-end
added in the follow-on CL).

There's a number of limitations this query module has today:

- Requires Python 3.8+, which means it doesn't work in the chroot, but
  works outside just fine (hope to have that resolved shortly).
- Overlays don't yet expose the portage masters.
- No query implementation for eclasses yet (but we can add that
  later).
- USE_EXPAND isn't yet expanded into USE flags.
- No ebuild dependencies yet.
- We could totally add more useful fields to lots of types.

But hey, it's a start.  We can do more later.

BUG=b:277149457
TEST=unit tests

Change-Id: I34690f4c97160bd2b8748518f335c473f8308521
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4405085
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>