scripts: cros_gdb - add option to pass gdb binary
It can be helpful to pass the gdb binary (especially in cros chrome_sdk)
especially when you want to have the option to parse python scripts in
gdb so that complex types can be displayed. Default is to use the old
behaviour only if cross_gdb is passed the binary from there will be
used.

BUG=b:278247253
TEST=cros_gdb --board $SDK_BOARD --remote dut_vm --attach browser
TEST=cros_gdb --board $SDK_BOARD --remote dut_vm --attach browser --cross_gdb /usr/bin/gdb

Change-Id: Ib4ced710efce70ad64cf899920d061ba502a164b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4428031
Reviewed-by: Tim Bain <tbain@google.com>
Tested-by: Marc Grimme <marcgrimme@google.com>
Commit-Queue: Marc Grimme <marcgrimme@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>