cros-ide: Refactors the code to compute shifts to be applied to threads
Previously, the code computed the local git diff and the shift computed
from the diff in a single loop. This made it difficult to further
refactor it.

The new code computes the local diff in chunks and then uses it to
compute the shift. This makes the code easier to refactor.

This is a pure refactoring change.

BUG=b:270101085
TEST=npm run unit-test

Change-Id: Iafc80d6a22c448c0f8f78d31d762714def178350
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4428914
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>