scripts: Add a script to generate "cros query" docs
Using the command line for "cros query" involves knowing the
attributes which can be queried.  Create a script which generates
Markdown documentation for each query target, and add tests that the
generated Markdown file matches.

This isn't meant to be complete documentation to "cros query", but
only the auto-generated side of things.  I plan to add a
docs/cros-query.md later with the higher-level discussion of the tool,
and link out to these docs as required.

BUG=b:277149457
TEST=unit tests
TEST=view markdown in gitiles

Change-Id: I4247063c95593d97f47d628ff319ed2f0965903b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4435355
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>