cros-ide: Add settings for boilerplate insertion
I left the setting for namespace guessing off by default, because I
noticed that it is not that good at the moment. One of the main
reasons is that `vscode.workspace.fs.readFile` appears to be super
slow (200ms per file!). I'll have to investigate this further at a
later stage.

BUG=b:274425473
TEST=Check that `cros-ide.boilerplate.enabled` can be used to disable
TEST=boilerplate insertion, and that
TEST=`cros-ide.boilerplate.guessNamespace` can be used to enable
TEST=namespace guessing.

Change-Id: Ibb0c090b09d63c5d8ebe3b7a3fe2c11d35a694cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4439909
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Christian Flach <cmfcmf@chromium.org>