lib: build_query: Correct incremental variable unsetting behavior
Previously, this used .update() from the resolved incremental
variables from the parent profile, which would not be correct when a
parent profile unsets a variable previously set by another parent.

For example:

   profile-a: USE="some another", parents=[]
   profile-b: USE="-another", parents=[]
   profile-c: USE="", parents=[profile-a, profile-b]

This should evaluate to USE="some", not USE="some another".

I added a test case similar to this in the unit tests.

BUG=b:277149457
TEST=unit tests

Change-Id: Icae28cd2b8ae3fb9584267de361b01a0ca99dbd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4452032
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>