lib: build_query: Implement .parents on overlays
Implement .parents on overlays, which accesses the Portage
masters (notice the COIL rename).  This makes the .tree() method on
overlays more understandable, and enables us to integrate the
make.conf variables into the board USE flags in a follow-up CL.

This required fixing a minor bug in portage_testables: when given a
list of parent overlays, it would not have a space between
eclass-overlay and the first parent overlay.

BUG=b:277149457
TEST=unit tests
TEST="cros query overlays -t" looks good

Change-Id: Ib8d62b5f261709e58220d9be27c06015f92d6e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4450782
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>