cli: cros_query: Fix access of globals in filter expressions
The contract of __getitem__ is to raise a KeyError instead of an
AttributeError when an item does not exist.

This prevents accessing the globals during expression evaluation, as
locals are evaluated before globals, and the non-existent attribute
raises an exception which is not caught.

BUG=b:277149457
TEST=unit tests

Change-Id: I761be25bb96bd1f637dca6170a5de3deb1efb385
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4454912
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>