api: controller: Fix default out_path
ParseChroot() was supposed to use DEFAULT_OUT_PATH if no out_path was
provided in the proto; but instead, we ended up with a default Path("")
-- i.e., the current directory.

This wasn't caught by unit tests, because chroot_lib.Chroot's equality
implementation was incomplete. Fix that at the same time, as well as a
cbuildbot test that becomes exposed.

BUG=b:265885353
TEST=./run_tests

Change-Id: I3d9e1fc75b9a0ac5b8c3fa421a2035ac6a6f99b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4477624
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Brian Norris <briannorris@chromium.org>