lib: build_query: Add is_variant property to boards
It can be useful to determine if a board is a "variant" of
another (note: the term "variant" is not being used here to refer to
the historical "variant-" overlays).  We define a variant board as a
board which has a parent overlay whose board name is not the same.

BUG=b:277149457
TEST=unit tests

Change-Id: Icd7755565b32cec237bde13a9f0f8f738560cb2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4500787
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>